add friends sidebar in every pages
