+++ front/app/src/js/components/buttons/index.js
@@ -1,5 +1,6 @@
+import {FriendsButton} from './FriendsButton.js';
 import {GithubButton} from './GithubButton.js';
 import {IntraButton} from './IntraButton.js';
 import {ThemeButton} from './ThemeButton.js';
 
-export {GithubButton, IntraButton, ThemeButton};
+export {FriendsButton, GithubButton, IntraButton, ThemeButton};
