+++ front/app/src/js/components/index.js
@@ -1,10 +1,16 @@
 import {Component} from './Component.js';
 
 import {
+  FriendsButton,
   GithubButton,
   IntraButton,
   ThemeButton,
 } from './buttons';
+
+import {
+  Friends,
+  FriendsSidebar,
+} from './layouts';
 
 import {
   ConnectedNavbar,
@@ -23,26 +29,32 @@ import {
   SignIn,
   SignUp,
   TwoFactorAuth,
+  Home,
+  HomeContent,
+  Local,
+  LocalContent,
+  Multiplayer,
+  MultiplayerContent,
   ResetPassword,
   ResetPasswordCode,
   ResetPasswordEmail,
   ResetPasswordNew,
   TournamentBracket,
   TournamentCreate,
+  TournamentCreateContent,
   TournamentDetails,
   Tournaments,
+  TournamentsContent,
   TournamentsList,
   UserProfile,
   UserProfileChart,
   UserProfileChartsCards,
+  UserProfileContent,
   UserProfileHeader,
   UserProfileMatchList,
   UserProfileStatsCard,
   UserProfileStatsCards,
-  Home,
-  Multiplayer,
   NotFound,
-  Singleplayer,
 } from './pages';
 
 import {
@@ -54,10 +66,13 @@ import {
   Game,
 } from './game';
 
-
+customElements.define('friends-button-component', FriendsButton);
 customElements.define('github-button-component', GithubButton);
 customElements.define('intra-button-component', IntraButton);
 customElements.define('theme-button-component', ThemeButton);
+
+customElements.define('friends-component', Friends);
+customElements.define('friends-sidebar-component', FriendsSidebar);
 
 customElements.define('connected-navbar-component', ConnectedNavbar);
 customElements.define('disconnected-navbar-component', DisconnectedNavbar);
@@ -72,6 +87,15 @@ customElements.define('two-factor-auth-component', TwoFactorAuth);
 customElements.define('signin-component', SignIn);
 customElements.define('signup-component', SignUp);
 
+customElements.define('home-component', Home);
+customElements.define('home-content-component', HomeContent);
+
+customElements.define('local-component', Local);
+customElements.define('local-content-component', LocalContent);
+
+customElements.define('multiplayer-component', Multiplayer);
+customElements.define('multiplayer-content-component', MultiplayerContent);
+
 customElements.define('reset-password-component', ResetPassword);
 customElements.define('reset-password-code-component', ResetPasswordCode);
 customElements.define('reset-password-email-component', ResetPasswordEmail);
@@ -79,19 +103,23 @@ customElements.define('reset-password-new-component', ResetPasswordNew);
 
 customElements.define('tournament-bracket-component', TournamentBracket);
 customElements.define('tournament-create-component', TournamentCreate);
+customElements.define(
+    'tournament-create-content-component', TournamentCreateContent,
+);
 customElements.define('tournament-details-component', TournamentDetails);
-customElements.define('tournaments-list-component', TournamentsList);
 customElements.define('tournaments-component', Tournaments);
+customElements.define('tournaments-content-component', TournamentsContent);
+customElements.define('tournaments-list-component', TournamentsList);
 
-customElements.define('home-component', Home);
-customElements.define('multiplayer-component', Multiplayer);
 customElements.define('notfound-component', NotFound);
-customElements.define('singleplayer-component', Singleplayer);
 
 customElements.define('user-profile-component', UserProfile);
 customElements.define('user-profile-chart-component', UserProfileChart);
 customElements.define(
     'user-profile-charts-cards-component', UserProfileChartsCards,
+);
+customElements.define(
+    'user-profile-content-component', UserProfileContent,
 );
 customElements.define('user-profile-header-component', UserProfileHeader);
 customElements.define(
@@ -112,24 +140,32 @@ customElements.define('game-component', Game);
 
 export {
   Component,
+  FriendsButton,
   GithubButton,
   IntraButton,
   ThemeButton,
+  Friends,
+  FriendsSidebar,
   SignIn,
   SignUp,
   TwoFactorAuth,
+  Local,
+  LocalContent,
   ResetPassword,
   ResetPasswordCode,
   ResetPasswordEmail,
   ResetPasswordNew,
   TournamentBracket,
   TournamentCreate,
+  TournamentCreateContent,
   TournamentDetails,
   Tournaments,
+  TournamentsContent,
   TournamentsList,
   UserProfile,
   UserProfileChart,
   UserProfileChartsCards,
+  UserProfileContent,
   UserProfileHeader,
   UserProfileMatchList,
   UserProfileStatsCard,
@@ -137,7 +173,6 @@ export {
   Home,
   Multiplayer,
   NotFound,
-  Singleplayer,
   Alert,
   Error,
   Navbar,
