+++ front/app/src/js/components/layouts/index.js
@@ -1,0 +1,4 @@
+import {FriendsSidebar} from './FriendsSidebar.js';
+import {Friends} from './Friends.js';
+
+export {FriendsSidebar, Friends};
