+++ front/app/src/js/components/pages/home/HomeContent.js
@@ -1,0 +1,19 @@
+import {Component} from '@components';
+
+export class HomeContent extends Component {
+  constructor() {
+    super();
+  }
+  render() {
+    return (`
+     <h1>Home</h1> 
+    `);
+  }
+  style() {
+    return (`
+      <style>
+
+      </style>
+    `);
+  }
+}
