+++ front/app/src/js/components/pages/home/index.js
@@ -1,0 +1,7 @@
+import {Home} from './Home.js';
+import {HomeContent} from './HomeContent.js';
+
+export {
+  Home,
+  HomeContent,
+};
