+++ front/app/src/js/components/pages/local/Local.js
@@ -1,13 +1,13 @@
 import {Component} from '@components';
 
-export class Singleplayer extends Component {
+export class Local extends Component {
   constructor() {
     super();
   }
   render() {
     return (`
       <navbar-component nav-active="singleplayer"></navbar-component>
-      <h1>Singleplayer</h1>
+      <friends-sidebar-component main-component="local-content-component"></friends-sidebar-component>
     `);
   }
   style() {
