+++ front/app/src/js/components/pages/local/index.js
@@ -1,0 +1,7 @@
+import {Local} from './Local.js';
+import {LocalContent} from './LocalContent.js';
+
+export {
+  Local,
+  LocalContent,
+};
