+++ front/app/src/js/components/pages/multiplayer/index.js
@@ -1,0 +1,7 @@
+import {Multiplayer} from './Multiplayer.js';
+import {MultiplayerContent} from './MultiplayerContent.js';
+
+export {
+  Multiplayer,
+  MultiplayerContent,
+};
