+++ front/app/src/js/components/pages/tournaments/index.js
@@ -1,13 +1,17 @@
 import {TournamentBracket} from './TournamentBracket.js';
 import {TournamentCreate} from './TournamentCreate.js';
+import {TournamentCreateContent} from './TournamentCreateContent.js';
 import {TournamentDetails} from './TournamentDetails.js';
 import {Tournaments} from './Tournaments.js';
+import {TournamentsContent} from './TournamentsContent.js';
 import {TournamentsList} from './TournamentsList.js';
 
 export {
   TournamentBracket,
   TournamentCreate,
+  TournamentCreateContent,
   TournamentDetails,
   Tournaments,
+  TournamentsContent,
   TournamentsList,
 };
