+++ front/app/src/js/components/pages/user_profile/index.js
@@ -1,6 +1,7 @@
 import {UserProfile} from './UserProfile.js';
 import {UserProfileChart} from './UserProfileChart.js';
 import {UserProfileChartsCards} from './UserProfileChartsCards.js';
+import {UserProfileContent} from './UserProfileContent.js';
 import {UserProfileHeader} from './UserProfileHeader.js';
 import {UserProfileMatchList} from './UserProfileMatchList.js';
 import {UserProfileStatsCard} from './UserProfileStatsCard.js';
@@ -10,6 +11,7 @@ export {
   UserProfile,
   UserProfileChart,
   UserProfileChartsCards,
+  UserProfileContent,
   UserProfileHeader,
   UserProfileMatchList,
   UserProfileStatsCard,
