+++ front/app/src/js/utils/cache/index.js
@@ -1,0 +1,4 @@
+import {Cache} from './Cache.js';
+import {FriendsCache} from './FriendsCache.js';
+
+export {Cache, FriendsCache};
